+++ GothicModComposer.UI/Converters/BooleanToVisibilityConverter.cs
@@ -1,0 +1,11 @@
+﻿using System.Windows;
+
+namespace GothicModComposer.UI.Converters
+{
+    public class BooleanToVisibilityConverter : BooleanConverter<Visibility>
+    {
+        public BooleanToVisibilityConverter() :
+            base(Visibility.Visible, Visibility.Collapsed)
+        { }
+    }
+}